+++ EstimoteFleetManagementSDK/EstimoteFleetManagementSDK.framework/Versions/A/Headers/ESTBeaconSettingsManager.h
@@ -1,4 +1,4 @@
-//  FleetManagementSDK
+//  Estimote Fleet Management SDK
 //  Copyright © 2015 Estimote. All rights reserved.
 
 #import <Foundation/Foundation.h>
@@ -26,7 +26,7 @@ NS_ASSUME_NONNULL_BEGIN
  */
 #define ESTDeviceSettingsManagerErrorDomain @"ESTDeviceSettingsManagerErrorDomain"
 
-typedef void(^ESTDeviceSettingsManagerSyncCompletionBlock)(NSError * _Nullable error);
+typedef void (^ESTDeviceSettingsManagerSyncCompletionBlock)(NSError *_Nullable error);
 
 /**
  *  Block used as a result of all operation success or failure of one of them.
@@ -35,13 +35,12 @@ typedef void(^ESTDeviceSettingsManagerSyncCompletionBlock)(NSError * _Nullable e
  *
  *  @param error Error of operation. No error means success.
  */
-typedef void(^ESTDeviceSettingsManagerOperationsCompletionBlock)(NSError * _Nullable error);
+typedef void (^ESTDeviceSettingsManagerOperationsCompletionBlock)(NSError *_Nullable error);
 
 /**
  *  Settings related errors.
  */
-typedef NS_ENUM(NSInteger, ESTDeviceSettingsManagerError)
-{
+typedef NS_ENUM (NSInteger, ESTDeviceSettingsManagerError) {
     /**
      *  Settings synchronization is already in progress for the device.
      */
@@ -58,18 +57,22 @@ typedef NS_ENUM(NSInteger, ESTDeviceSettingsManagerError)
      *  Validation failed for another setting provided in the group.
      */
     ESTDeviceSettingsManagerErrorSettingValidationFailed,
-    
+
     /**
      *  Reading setting value from Estimote Cloud failed.
      */
     ESTDeviceSettingsManagerErrorSettingCloudReadFailed,
-    
+
     /**
      *  Saving new value of settings in Estimote Cloud failed.
      */
-    ESTDeviceSettingsManagerErrorSettingCloudSaveFailed
-};
+    ESTDeviceSettingsManagerErrorSettingCloudSaveFailed,
 
+    /**
+     *  Time synchronization with Estimote Cloud failed.
+     */
+    ESTDeviceSettingsManagerErrorTimeSyncWithCloudFailed
+};
 
 /**
  *  ESTDeviceSettingsManager allows to manager device settings interaction
@@ -152,7 +155,6 @@ typedef NS_ENUM(NSInteger, ESTDeviceSettingsManagerError)
  */
 @property (nonatomic, strong, readonly) ESTSettingsEddystoneConfigurationService *eddystoneConfigurationService;
 
-
 #pragma mark - Advanced Settings API
 ///--------------------------------------------------------------------
 /// @name Advanced Settings API
@@ -177,7 +179,7 @@ typedef NS_ENUM(NSInteger, ESTDeviceSettingsManagerError)
  *
  *  @param operationsArray Array of setting operations.
  */
-- (void)performOperationsFromArray:(NSArray<id<ESTBeaconOperationProtocol>> *)operationsArray;
+- (void)performOperationsFromArray:(NSArray<id<ESTBeaconOperationProtocol> > *)operationsArray;
 
 /**
  *  Method allows to perform group of setting read/write operations provided in an array.
@@ -185,7 +187,7 @@ typedef NS_ENUM(NSInteger, ESTDeviceSettingsManagerError)
  *  @param operationsArray Array of setting operations.
  *  @param completion Completion block that should be executed after all operations succeed or any of the operation fails.
  */
-- (void)performOperationsFromArray:(NSArray<id<ESTBeaconOperationProtocol>> *)operationsArray completion:(ESTDeviceSettingsManagerOperationsCompletionBlock _Nullable)completion;
+- (void)performOperationsFromArray:(NSArray<id<ESTBeaconOperationProtocol> > *)operationsArray completion:(ESTDeviceSettingsManagerOperationsCompletionBlock _Nullable)completion;
 
 #pragma mark Notifications
 ///--------------------------------------------------------------------
@@ -205,7 +207,6 @@ typedef NS_ENUM(NSInteger, ESTDeviceSettingsManagerError)
 - (void)unregisterAllNotifications;
 
 @end
-
 
 @class ESTDeviceLocationBeacon;